Reduce code duplication and only use one manageImportJob method
